refactor(infrastructure): extract AppDbContext helpers into dedicated classes (#277)
- Extract TypedIdConventionApplier with LINQ-filtered reflection for Entity/Property/HasConversion methods
- Extract SoftDeleteFilterApplier to manage query filters and timestamp configuration
- Extract RoleSeedDataConfiguration as internal static with 10 seed ID constants and HasData blocks
- Create AssemblyInfo.cs with InternalsVisibleTo for DataSeeder and test assemblies
- Update AppDbContext from 874 to 596 lines with clean delegation to helpers
- Update 26 external seed ID call sites across 11 files (unchanged GUID values)
- All 1,236 tests pass (DataSeeder 38, Architecture 42, Unit 758, Integration 398)
- Zero EF migration diff (model snapshot unchanged)
