+++ LgymApi.Infrastructure/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("LgymApi.DataSeeder")]
+[assembly: InternalsVisibleTo("LgymApi.DataSeeder.Tests")]
+[assembly: InternalsVisibleTo("LgymApi.TestUtils")]
+[assembly: InternalsVisibleTo("LgymApi.IntegrationTests")]
