+++ LgymApi.IntegrationTests/Pagination/TrainerDashboardGridifyPaginationTests.cs
@@ -7,6 +7,7 @@ using LgymApi.Domain.Entities;
 using LgymApi.Domain.Enums;
 using LgymApi.Domain.ValueObjects;
 using LgymApi.Infrastructure.Data;
+using LgymApi.Infrastructure.Data.SeedData;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
@@ -111,13 +112,13 @@ public sealed class TrainerDashboardGridifyPaginationTests : IntegrationTestBase
         using var scope = Factory.Services.CreateScope();
         var db = scope.ServiceProvider.GetRequiredService<AppDbContext>();
 
-        var alreadyLinked = await db.UserRoles.AnyAsync(ur => ur.UserId == trainer.Id && ur.RoleId == AppDbContext.TrainerRoleSeedId);
+        var alreadyLinked = await db.UserRoles.AnyAsync(ur => ur.UserId == trainer.Id && ur.RoleId == RoleSeedDataConfiguration.TrainerRoleSeedId);
         if (!alreadyLinked)
         {
             db.UserRoles.Add(new UserRole
             {
                 UserId = trainer.Id,
-                RoleId = AppDbContext.TrainerRoleSeedId
+                RoleId = RoleSeedDataConfiguration.TrainerRoleSeedId
             });
         }
