+++ LgymApi.IntegrationTests/ReportingEngineTests.cs
@@ -6,6 +6,7 @@ using FluentAssertions;
 using LgymApi.Domain.Entities;
 using LgymApi.Domain.ValueObjects;
 using LgymApi.Infrastructure.Data;
+using LgymApi.Infrastructure.Data.SeedData;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -204,13 +205,13 @@ public sealed class ReportingEngineTests : IntegrationTestBase
 
          using var scope = Factory.Services.CreateScope();
          var db = scope.ServiceProvider.GetRequiredService<AppDbContext>();
-         var alreadyLinked = await db.UserRoles.AnyAsync(ur => ur.UserId == (Domain.ValueObjects.Id<User>)trainer.Id && ur.RoleId == (Domain.ValueObjects.Id<Role>)AppDbContext.TrainerRoleSeedId);
+         var alreadyLinked = await db.UserRoles.AnyAsync(ur => ur.UserId == (Domain.ValueObjects.Id<User>)trainer.Id && ur.RoleId == (Domain.ValueObjects.Id<Role>)RoleSeedDataConfiguration.TrainerRoleSeedId);
          if (!alreadyLinked)
          {
              db.UserRoles.Add(new UserRole
              {
                  UserId = (Domain.ValueObjects.Id<User>)trainer.Id,
-                 RoleId = (Domain.ValueObjects.Id<Role>)AppDbContext.TrainerRoleSeedId
+                 RoleId = (Domain.ValueObjects.Id<Role>)RoleSeedDataConfiguration.TrainerRoleSeedId
              });
              await db.SaveChangesAsync();
          }
