+++ LgymApi.IntegrationTests/SupplementationApiTests.cs
@@ -5,6 +5,7 @@ using FluentAssertions;
 using LgymApi.Domain.Entities;
 using LgymApi.Domain.ValueObjects;
 using LgymApi.Infrastructure.Data;
+using LgymApi.Infrastructure.Data.SeedData;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -235,13 +236,13 @@ public sealed class SupplementationApiTests : IntegrationTestBase
 
         using var scope = Factory.Services.CreateScope();
          var db = scope.ServiceProvider.GetRequiredService<AppDbContext>();
-         var alreadyLinked = await db.UserRoles.AnyAsync(ur => ur.UserId == trainer.Id && ur.RoleId == AppDbContext.TrainerRoleSeedId);
+         var alreadyLinked = await db.UserRoles.AnyAsync(ur => ur.UserId == trainer.Id && ur.RoleId == RoleSeedDataConfiguration.TrainerRoleSeedId);
          if (!alreadyLinked)
          {
               db.UserRoles.Add(new UserRole
               {
                   UserId = trainer.Id,
-                  RoleId = AppDbContext.TrainerRoleSeedId
+                  RoleId = RoleSeedDataConfiguration.TrainerRoleSeedId
               });
              await db.SaveChangesAsync();
          }
