+++ LgymApi.IntegrationTests/TrainerRelationshipTests.cs
@@ -9,6 +9,7 @@ using LgymApi.Domain.Entities;
 using LgymApi.Domain.Enums;
 using LgymApi.Domain.ValueObjects;
 using LgymApi.Infrastructure.Data;
+using LgymApi.Infrastructure.Data.SeedData;
 using LgymApi.Resources;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
@@ -1246,13 +1247,13 @@ public sealed class TrainerRelationshipTests : IntegrationTestBase
          using var scope = Factory.Services.CreateScope();
          var db = scope.ServiceProvider.GetRequiredService<AppDbContext>();
 
-          var alreadyLinked = await db.UserRoles.AnyAsync(ur => ur.UserId == trainer.Id && ur.RoleId == AppDbContext.TrainerRoleSeedId);
+          var alreadyLinked = await db.UserRoles.AnyAsync(ur => ur.UserId == trainer.Id && ur.RoleId == RoleSeedDataConfiguration.TrainerRoleSeedId);
          if (!alreadyLinked)
          {
               db.UserRoles.Add(new UserRole
               {
                   UserId = trainer.Id,
-                  RoleId = AppDbContext.TrainerRoleSeedId
+                  RoleId = RoleSeedDataConfiguration.TrainerRoleSeedId
              });
          }
